# openapi_spec.yaml
components:
  schemas:
...
    Body_import_hazards_csv_api_v1_operator__tenant_id__hazards_import_post:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
      required:
      - file
      title: Body_import_hazards_csv_api_v1_operator__tenant_id__hazards_import_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    Hazard:
      properties:
        category:
          default: ''
          title: Category
          type: string
        closed_at:
          default: ''
          title: Closed At
          type: string
        created_at:
          title: Created At
          type: string
        created_by:
          default: ''
          title: Created By
          type: string
        description:
          title: Description
          type: string
        hazard_id:
          default: ''
          title: Hazard Id
          type: string
        id:
          default: ''
          title: Id
          type: string
        source_report_ids:
          default: []
          items:
            type: string
          title: Source Report Ids
          type: array
        status:
          default: open
          title: Status
          type: string
        tenant_id:
          default: ''
          title: Tenant Id
          type: string
        title:
          title: Title
          type: string
      required:
      - title
      - description
      title: Hazard
      type: object
...
    RiskAssessment:
      properties:
        assessed_at:
          title: Assessed At
          type: string
        assessed_by:
          default: ''
          title: Assessed By
          type: string
        hazard_id:
          default: ''
          title: Hazard Id
          type: string
        id:
          default: ''
          title: Id
          type: string
        likelihood:
          default: 0
          title: Likelihood
          type: integer
        rationale:
          default: ''
          title: Rationale
          type: string
        risk_index:
          default: ''
          title: Risk Index
          type: string
        risk_zone:
          default: ''
          title: Risk Zone
          type: string
        severity:
          default: ''
          title: Severity
          type: string
      title: RiskAssessment
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: "Multi-tenant SMS aligned with ICAO Annex 19 / Doc 9859\n\nOperators:\n  - Sita Air (tenant_id: sita-air, API Key: sita-air-key-2026)\n\nRegulators:\n  - CAAN — Civil Aviation Authority of Nepal (tenant_id: caan, API Key: caan-reg-key-2026)"
  title: AviaSafe — Safety Management System
  version: 3.0.0
openapi: 3.1.0
paths:
...
  /api/v1/operator/{tenant_id}/hazards:
    get:
      operationId: list_hazards_api_v1_operator__tenant_id__hazards_get
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      - in: query
        name: status
        required: false
        schema:
          default: ''
          title: Status
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Hazards
    post:
      operationId: create_hazard_api_v1_operator__tenant_id__hazards_post
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hazard'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Hazard
  /api/v1/operator/{tenant_id}/hazards/import:
    post:
      operationId: import_hazards_csv_api_v1_operator__tenant_id__hazards_import_post
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_hazards_csv_api_v1_operator__tenant_id__hazards_import_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Import Hazards Csv
  /api/v1/operator/{tenant_id}/hazards/{hazard_id}:
    get:
      operationId: get_hazard_api_v1_operator__tenant_id__hazards__hazard_id__get
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      - in: path
        name: hazard_id
        required: true
        schema:
          title: Hazard Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Hazard
...
  /api/v1/operator/{tenant_id}/hazards/{hazard_id}/risk:
    post:
      operationId: assess_hazard_risk_api_v1_operator__tenant_id__hazards__hazard_id__risk_post
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      - in: path
        name: hazard_id
        required: true
        schema:
          title: Hazard Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskAssessment'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Assess Hazard Risk
...